+++ NEWS.md
@@ -1,3 +1,11 @@
+# huito 0.1.1
+
+- Update pkgdown
+- Update vignettes.
+- Include example from GerminaR sticker 
+- Match argument for `label_print(mode)` to "preview" and "sample".
+- Set the number of label to print: `label_print(nlabels)`
+
 # huito 0.1.0
 
 - Added a `NEWS.md` file to track changes to the package.